feat: policy rules crud methods (#36)
DRI reviewer: @ankushswar1 @jagtejsodhi 

https://app.asana.com/1/1202677445535453/project/1210977604891464/task/1211212943227768?focus=true
https://app.asana.com/1/1202677445535453/project/1210977604891464/task/1211212943227770?focus=true
https://app.asana.com/1/1202677445535453/project/1210977604891464/task/1211212943227772?focus=true


## Rules subresource

"Rules" are declared as a "subresource" of policies, which is semantically correct. However, stainless will codegenerate the `.rules` field under the `Policies` class, meaning the generated code will do the following:

```ts
export class Policies extends APIResource {
  rules: RulesAPI.Rules = new RulesAPI.Rules(this._client);
  // ...
}

privyClient.policies().rules.get(...)
```

This is problematic for a few reasons:

1) `rules` is a property (not a method) which breaks the pattern we had of `.policies()`, `.wallets()`, etc.
2) `rules` is a readonly property of type `Rules` -- meaning we cannot overload or much less change its type to our own `PrivyPolicyRulesService` subclass.

I've attempted to generate the subresource under the name `_rules` but failed to do so, coming against codegen errors. Because of this, I've opted for moving the methods up a level, all coexisting under the `PrivyPoliciesService` class. Still, `PrivyPolicyRulesService` exists internally and can be used eventually to expose under a `rules()` property when/if that becomes available.